wip develop mistake fix

+++ src/main/java/com/nobledigitalservice/noblecrm/asset_admin/model/AdminURLs.java
@@ -5,9 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
+import javax.persistence.*;
 
 @Entity
 @Data
@@ -22,8 +20,9 @@ public class AdminURLs {
     @Column
     private String url;
 
+
     @Column
-    private int id;
+    private Long id;
 
 
 }
